Ensure consistent ordering of network_mappings
The provider_networks module returned the network_mappings in a random
order changing with every invocation. This returns the entries sorted
and adds a test to ensure the ordering is consistent between
invocations.

This is the same change as commit
cccf4c03f134fbdcbe0664610064fcfc73671ad8 in openstack-ansible-plugins.
This change in the os_neutron role is required until Xena. Since Yoga
the version in openstack-ansible-plugins (aka openstack.osa collection)
is used.

This commit is not upstreamed as the relevant code is no longer present
in upstream master.